% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_egm_name.R
\name{clean_egm}
\alias{clean_egm}
\title{Convert Victorian gaming machine names to the latest values in a consistent format}
\usage{
clean_egm(
  x,
  to = "venue_abbr",
  fuzzy_match = TRUE,
  max_dist = 0.4,
  method = "jw"
)
}
\arguments{
\item{x}{a (character) vector containing Victorian gambling venue names}

\item{to}{what form should the state names be converted to? Options are
"venue_name" and "venue_abbr" (the default).}

\item{fuzzy_match}{logical; either TRUE (the default) which indicates that
approximate/fuzzy string matching should be used, or FALSE which indicates that
only exact matches should be used.}

\item{max_dist}{numeric, sets the maximum acceptable distance between your
string and the matched string. Default is 0.4. Only relevant when fuzzy_match is TRUE.}

\item{method}{the method used for approximate/fuzzy string matching. Default
is "jw", the Jaro-Winker distance; see \code{??stringdist-metrics} for more options.
#'}
}
\value{
a character vector of state names, abbreviations, or codes.
}
\description{
Convert Victorian gaming machine names to the latest values in a consistent format
}
