% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{egm_lga_data}
\alias{egm_lga_data}
\title{Victorian gaming venue data by LGA}
\format{
A data frame with 8100 rows and 5 variables:
\describe{
\item{\code{lga_name}}{The name of the local goverment area, cleaned with vpstheme::clean_vic_lga}
\item{\code{financial_year}}{The financial year of the data, as a character variable.}
\item{\code{data_month}}{The month of the data, as a Date variable.}
\item{\code{measure_type}}{The type of data stored in value, either 'EGM' being the number of Machines, 'Venues' being the number of Venues, or 'Expenditure' being player loss}
\item{\code{value}}{A numeric being the amount of measure_type for that LGA and month}

}
}
\source{
\url{https://www.vgccc.vic.gov.au/resources/information-and-data/expenditure-data}
}
\usage{
egm_lga_data
}
\description{
Tidy data set of Victorian gaming machine data by lga and month
}
\keyword{datasets}
